DOF to vertex map fix
The only significant change involves solving for the DOF to vertex map in the Step1 file. This value is then exported to a .txt file and read back in S1. There is also a significant amount of troubleshooting lines and comments that will be cleaned up in a future commit