Add process checks to save time
Added significant troubleshooting if statements to the nearest neighbor matching portion of the code. Moving forward with the pipeline, all values coming in should not need to be interpolated so nearest neighbor calcualtions need not take place in some circumstances. These improvements yeild a magnitude decrease in code run time in a benchmarking testing. Other naming of variables was done to help make the code more readable.