Save only linear tet nodes to new mesh
Previously all of the nodes from the original mesh were saved to the linear tet mesh files. Although the quadratic nodes should not have been referenced because they were removed from the connectivity matrix, they were removed to speed up the code and to ensure that the code functions properly. The files still display correctly and various other checks were conducted to ensure the files were not corrupted as a results of the change